eduvmstorebackend/requirements.txt
Django
djangorestframework
openstacksdk
sqlalchemy
ruff
vulture
bandit